Login, logout working
added tailwind and currectly redoing the entire front end using tailwind to make it more consistant. i was thinking of keeping bootstrap in the code so i didn't have to redo stuff but it ended up conflicting with tailwind so i'll just redo everything and try to make it as clean looking as possible.